refactor: seperate ProxyChallengeProcessorImpl into Basic and Digest two files with a new ProxyAuthenticator
