+++ src/main/java/com/microsoft/azure/proton/transport/proxy/ProxyChallengeProcessor.java
@@ -3,6 +3,6 @@ package com.microsoft.azure.proton.transport.proxy;
 import java.util.Map;
 
 public interface ProxyChallengeProcessor {
-    Map<String, String> getHeader(String challengeResp, String host);
+    Map<String, String> getHeader();
 }
