+++ src/main/java/com/microsoft/azure/proton/transport/proxy/impl/DigestProxyChallengeProcessorImpl.java
@@ -11,32 +11,33 @@ import java.security.SecureRandom;
 import java.util.*;
 import java.util.concurrent.atomic.AtomicInteger;
 
-public class ProxyChallengeProcessorImpl implements ProxyChallengeProcessor {
+public class DigestProxyChallengeProcessorImpl implements ProxyChallengeProcessor {
 
     private final String DIGEST = "digest";
-    private final String BASIC = "basic";
     private final String PROXY_AUTH_DIGEST = "Proxy-Authenticate: Digest";
-    private final String PROXY_AUTH_BASIC = "Proxy-Authenticate: Basic";
     private final AtomicInteger nonceCounter = new AtomicInteger(0);
-    private static Map<String, String> headers;
-    private static String host;
+    private final Map<String, String> headers;
+    private final ProxyAuthenticator proxyAuthenticator;
 
+    private static String host;
+    private static String challenge;
+
+    DigestProxyChallengeProcessorImpl(String host, String challenge) {
+        this.host = host;
+        this.challenge = challenge;
+        headers = new HashMap<>();
+        proxyAuthenticator = new ProxyAuthenticator();
+    }
 
     @Override
-    public Map<String, String> getHeader(String challengeResp,
-                                         String host) {
-        this.host = host;
-
-        final Scanner responseScanner = new Scanner(challengeResp);
-        final Map<String, String> challengeQuestionValues = new HashMap<String, String>();
+    public Map<String, String> getHeader() {
+        final Scanner responseScanner = new Scanner(challenge);
+        final Map<String, String> challengeQuestionValues = new HashMap<>();
         while (responseScanner.hasNextLine()) {
             String line = responseScanner.nextLine();
             if (line.contains(PROXY_AUTH_DIGEST)){
                 getChallengeQuestionHeaders(line, challengeQuestionValues);
-                computeDigestAuthHeader(challengeQuestionValues, host, getPasswordAuthentication(DIGEST));
-                break;
-            } else if (line.contains(PROXY_AUTH_BASIC)) {
-                computeBasicAuthHeader(getPasswordAuthentication(BASIC));
+                computeDigestAuthHeader(challengeQuestionValues, host, proxyAuthenticator.getPasswordAuthentication(DIGEST, host));
                 break;
             }
         }
@@ -56,52 +57,10 @@ public class ProxyChallengeProcessorImpl implements ProxyChallengeProcessor {
         }
     }
 
-    private PasswordAuthentication getPasswordAuthentication(String scheme) {
-        ProxySelector proxySelector = ProxySelector.getDefault();
-
-        URI uri;
-        List<Proxy> proxies = null;
-        if (host != null && !host.isEmpty()) {
-            uri = URI.create(host);
-            proxies = proxySelector.select(uri);
-        }
-
-        InetAddress proxyAddr = null;
-        Proxy.Type proxyType = null;
-        if (isProxyAddressLegal(proxies)) {
-            // will be only one element in the proxy list
-            proxyAddr = ((InetSocketAddress)proxies.get(0).address()).getAddress();
-            proxyType = proxies.get(0).type();
-        }
-        return Authenticator.requestPasswordAuthentication(
-                "",
-                proxyAddr,
-                0,
-                proxyType == null ? "" : proxyType.name(),
-                "Event Hubs client websocket proxy support",
-                scheme,
-                null,
-                Authenticator.RequestorType.PROXY);
-    }
-
-    private void computeBasicAuthHeader(PasswordAuthentication passwordAuthentication ){
-        if (!isPasswordAuthenticationHasValues(passwordAuthentication))
-            return;
-
-        String proxyUserName = passwordAuthentication.getUserName();
-        String proxyPassword = new String(passwordAuthentication.getPassword());
-        final String usernamePasswordPair = proxyUserName + ":" + proxyPassword;
-        if (headers == null)
-            headers = new HashMap<String, String>();
-        headers.put(
-                "Proxy-Authorization",
-                "Basic " + Base64.getEncoder().encodeToString(usernamePasswordPair.getBytes()));
-    }
-
     private void computeDigestAuthHeader(Map<String, String> challengeQuestionValues,
                                          String uri,
                                          PasswordAuthentication passwordAuthentication) {
-        if (!isPasswordAuthenticationHasValues(passwordAuthentication))
+        if (!proxyAuthenticator.isPasswordAuthenticationHasValues(passwordAuthentication))
             return;
 
         String proxyUserName = passwordAuthentication.getUserName();
@@ -132,39 +91,11 @@ public class ProxyChallengeProcessorImpl implements ProxyChallengeProcessor {
                         proxyUserName, realm, nonce, uri, cnonce, nc, response, qop);
             }
 
-            if (headers == null) {
-                headers = new HashMap<>();
-            }
             headers.put("Proxy-Authorization", digestValue);
-
         } catch(NoSuchAlgorithmException ex) {
             throw new RuntimeException(ex);
         } catch (UnsupportedEncodingException ex) {
             throw new RuntimeException(ex);
         }
-    }
-
-
-    private boolean isNullOrEmpty(String string) {
-        return (string == null || string.isEmpty());
-    }
-
-    private boolean isPasswordAuthenticationHasValues(PasswordAuthentication passwordAuthentication){
-        if (passwordAuthentication == null) return false;
-        String proxyUserName = passwordAuthentication.getUserName() != null
-                ? passwordAuthentication.getUserName()
-                : null ;
-        String proxyPassword = passwordAuthentication.getPassword() != null
-                ? new String(passwordAuthentication.getPassword())
-                : null;
-        if (isNullOrEmpty(proxyUserName) || isNullOrEmpty(proxyPassword))  return false;
-        return true;
-    }
-
-    private static boolean isProxyAddressLegal(final List<Proxy> proxies) {
-        return proxies != null
-                && !proxies.isEmpty()
-                && proxies.get(0).address() != null
-                && proxies.get(0).address() instanceof InetSocketAddress;
     }
 }
